+++ src/gen/genModule.test.ts
@@ -1,5 +1,8 @@
+import {join} from 'path';
+
 import {test, t} from '../oki/oki.js';
-import {validateGenModule} from './genModule.js';
+import {validateGenModule, findGenModules} from './genModule.js';
+import {paths} from '../paths.js';
 
 test('validateGenModule()', () => {
 	t.ok(validateGenModule({gen: () => {}}));
@@ -8,4 +11,9 @@ test('validateGenModule()', () => {
 		t.ok(!validateGenModule({gen: {}}));
 		t.ok(!validateGenModule({task: {run: {}}}));
 	});
+});
+
+test('findGenModules()', async () => {
+	const findGenModulesResult = await findGenModules([join(paths.source, 'docs/')]);
+	t.ok(findGenModulesResult.ok);
 });